+++ packages/react-stream/inline.dom.js
@@ -8,4 +8,4 @@
 // This file intentionally does *not* have the Flow annotation.
 // Don't add it. See `./inline-typed.js` for an explanation.
 
-export * from './src/ReactFizzRenderer';
+export * from './src/ReactFizzStreamer';
